+++ src/FolkerKinzel.Strings/Intls/Base64.cs
@@ -1,9 +1,4 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Diagnostics;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
+﻿using System.Diagnostics;
 using FolkerKinzel.Strings.Polyfills;
 
 namespace FolkerKinzel.Strings.Intls;
@@ -11,7 +6,7 @@ namespace FolkerKinzel.Strings.Intls;
 [SuppressMessage("Globalization", "CA1303:Literale nicht als lokalisierte Parameter übergeben", Justification = "<Ausstehend>")]
 internal static class Base64
 {
-    internal const string LINE_BREAK = "\n";
+    internal const string LINE_BREAK = "\r\n";
     internal const int LINE_LENGTH = 76;
 
     private const string IDX = "ABCDEFGHIJKLMNOPQRSTUVWXYZabcdefghijklmnopqrstuvwxyz0123456789+/";
@@ -19,13 +14,14 @@ internal static class Base64
     private const int CHUNK_LENGTH = 3;
     private const int CHAR_WIDTH = 6;
 
-    internal static void AppendEncodedTo(StringBuilder sb, IList<byte> data)
+    internal static void AppendEncodedTo(StringBuilder sb, ReadOnlySpan<byte> data)
     {
         Debug.Assert(sb != null);
-        Debug.Assert(data != null);
 
-        int paddingLength = CHUNK_LENGTH - (data.Count % CHUNK_LENGTH);
-        int finalPartLength = paddingLength == 0 ? 0 : CHUNK_LENGTH - paddingLength;
+        // paddingLength may be 3, but finalPartLength is 0 then and no 
+        // padding will be added:
+        int paddingLength = CHUNK_LENGTH - (data.Length % CHUNK_LENGTH);
+        int finalPartLength = CHUNK_LENGTH - paddingLength;
 
         AppendChunks(sb, data, finalPartLength);
 
@@ -35,9 +31,9 @@ internal static class Base64
         }
     }
 
-    private static void AppendChunks(StringBuilder sb, IList<byte> data, int finalPartLength)
+    private static void AppendChunks(StringBuilder sb, ReadOnlySpan<byte> data, int finalPartLength)
     {
-        if (data.Count < CHUNK_LENGTH)
+        if (data.Length < CHUNK_LENGTH)
         {
             return;
         }
@@ -46,7 +42,7 @@ internal static class Base64
         ReadOnlySpan<char> idx = IDX.AsSpan();
         Span<char> tmp = stackalloc char[4];
 
-        while (counter < data.Count - finalPartLength)
+        while (counter < data.Length - finalPartLength)
         {
             int i = data[counter] << 16;
             i |= data[counter + 1] << 8;
@@ -64,7 +60,7 @@ internal static class Base64
     }
 
 
-    private static void AppendFinalBlock(StringBuilder sb, IList<byte> data, int paddingLength, int finalPartLength)
+    private static void AppendFinalBlock(StringBuilder sb, ReadOnlySpan<byte> data, int paddingLength, int finalPartLength)
     {
         ReadOnlySpan<char> idx = IDX.AsSpan();
         int dataHolder = 0;
@@ -72,7 +68,7 @@ internal static class Base64
         for (int j = 0; j < finalPartLength; j++)
         {
             dataHolder <<= 8;
-            dataHolder |= data[data.Count - (finalPartLength - j)];
+            dataHolder |= data[data.Length - (finalPartLength - j)];
         }
 
         dataHolder <<= paddingLength * 2;
@@ -90,33 +86,4 @@ internal static class Base64
             sb.Append('=');
         }
     }
-
-    //private void InitIdx(Span<char> span)
-    //{
-    //    // Absolute offset for all ranges:
-    //    // Translate values 0..63 to the Base64 alphabet. There are five sets:
-    //    // #  From      To         Abs    Index  Characters
-    //    // 0  [0..25]   [65..90]   +65        0  ABCDEFGHIJKLMNOPQRSTUVWXYZ
-    //    // 1  [26..51]  [97..122]  +71        1  abcdefghijklmnopqrstuvwxyz
-    //    // 2  [52..61]  [48..57]    -4  [2..11]  0123456789
-    //    // 3  [62]      [43]       -19       12  +
-    //    // 4  [63]      [47]       -16       13  /
-    //    for (int i = 0; i < 26; i++)
-    //    {
-    //        span[i] = (char)(i + 65);
-    //    }
-
-    //    for (int i = 26; i < 52; i++)
-    //    {
-    //        span[i] = 71;
-    //    }
-
-    //    for (int i = 52; i < 62; i++)
-    //    {
-    //        span[i] = -4;
-    //    }
-
-    //    span[62] = -19;
-    //    span[63] = -16;
-    //}
 }