+++ src/FolkerKinzel.Strings/StringBuilderExtension_AppendBase64Encoded.cs
@@ -10,10 +10,11 @@ public static partial class StringBuilderExtension
                                                     IEnumerable<byte> bytes,
                                                     Base64FormattingOptions options = Base64FormattingOptions.None)
     {
-        if (builder is null) { throw new ArgumentNullException(nameof(builder)); }
-        if (bytes is null) { throw new ArgumentNullException(nameof(bytes)); }
+        if (bytes is null) 
+        { 
+            throw new ArgumentNullException(nameof(bytes)); 
+        }
 
-        int startOfBase64 = builder.Length;
         ReadOnlySpan<byte> span = bytes switch
         {
             byte[] arr => arr,
@@ -23,10 +24,24 @@ public static partial class StringBuilderExtension
             _ => bytes.ToArray(),
         };
 
-        bool insertLineBreaks = options.HasFlag(Base64FormattingOptions.InsertLineBreaks) && !span.IsEmpty;
-        builder.EnsureCapacity(builder.Length + ComputeNeededCapacity(span.Length, insertLineBreaks));
+        return builder.AppendBase64Encoded(span, options);
+    }
 
-        Base64New.AppendEncodedTo(builder, span);
+
+    public static StringBuilder AppendBase64Encoded(this StringBuilder builder,
+                                                    ReadOnlySpan<byte> bytes,
+                                                    Base64FormattingOptions options = Base64FormattingOptions.None)
+    {
+        if (builder is null) 
+        { 
+            throw new ArgumentNullException(nameof(builder));
+        }
+
+        bool insertLineBreaks = options.HasFlag(Base64FormattingOptions.InsertLineBreaks) && !bytes.IsEmpty;
+        builder.EnsureCapacity(builder.Length + ComputeNeededCapacity(bytes.Length, insertLineBreaks));
+        int startOfBase64 = builder.Length;
+
+        Base64.AppendEncodedTo(builder, bytes);
 
         if (insertLineBreaks)
         {
@@ -42,7 +57,7 @@ public static partial class StringBuilderExtension
 
         if (insertLineBreaks)
         {
-            capacity += dataLength / Base64New.LINE_LENGTH * Base64New.LINE_BREAK.Length;
+            capacity += dataLength / Base64.LINE_LENGTH * Base64.LINE_BREAK.Length;
         }
 
         return capacity;
@@ -50,18 +65,24 @@ public static partial class StringBuilderExtension
 
     private static void InsertLineBreaks(StringBuilder builder, int startOfBase64)
     {
-        int lastLineBreakIndex = builder.LastIndexOf('\n', startOfBase64);
-
-        if (lastLineBreakIndex >= 0)
-        {
-            if (startOfBase64 - lastLineBreakIndex < Base64New.LINE_LENGTH) { startOfBase64 = lastLineBreakIndex; }
+        int currentLineStartIndex = builder.LastIndexOf('\n', startOfBase64) + 1;
+        
+        if (startOfBase64 - currentLineStartIndex < Base64.LINE_LENGTH)
+        { 
+            startOfBase64 = currentLineStartIndex; 
         }
+        else
+        {
+            builder.Insert(startOfBase64, Base64.LINE_BREAK);
+            startOfBase64 += Base64.LINE_BREAK.Length;
+        }
+        
 
-        int nextLineStart = startOfBase64 + Base64New.LINE_LENGTH;
+        int nextLineStart = startOfBase64 + Base64.LINE_LENGTH;
         while (nextLineStart < builder.Length)
         {
-            builder.Insert(nextLineStart, Base64New.LINE_BREAK);
-            nextLineStart += Base64New.LINE_BREAK.Length + Base64New.LINE_LENGTH;
+            builder.Insert(nextLineStart, Base64.LINE_BREAK);
+            nextLineStart += Base64.LINE_BREAK.Length + Base64.LINE_LENGTH;
         }
     }